cloud-sync: add "await" in front of RPC calls
So async errors are caught correctly. Also, consolidate the two
copies of the same code into one module.

Fixes #617
